django-boilerplate/requirements.txt
asgiref==3.8.1
boto3==1.34.98
botocore==1.34.98
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
firebase-admin==6.5.0
gunicorn==22.0.0
googlemaps==4.10.0
jmespath==1.0.1
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
wget==3.2
whitenoise==6.6.0
drf-yasg==1.21.7
pyyaml==6.0.1
uritemplate==4.1.1
django-rest-passwordreset==1.4.1
django-debug-toolbar==4.3.0
django-import-export==4.0.0
timeago==1.0.16
celery==5.4.0
flower==2.0.1
redis==5.0.4
django-redis==5.4.0
pillow==10.3.0
pillow-heif==0.16.0
pymongo==4.7.1
qrcode
django-otp
openai==1.54.0